started seperating pages and server-actions, instead of having it all in pagelogin
